Move TenantsEndpoint from generic Endpoints to Tenant namespace
