+++ account-management/Tests/WebApi/Endpoints/TenantEndpointsTests.cs
@@ -9,7 +9,7 @@ using PlatformPlatform.AccountManagement.Application.Tenants.Dtos;
 using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using PlatformPlatform.AccountManagement.Tests.Infrastructure;
-using PlatformPlatform.AccountManagement.WebApi.Endpoints;
+using PlatformPlatform.AccountManagement.WebApi.Tenants;
 using PlatformPlatform.Foundation.DomainModeling.Validation;
 using Xunit;
