+++ account-management/WebApi/Program.cs
@@ -1,7 +1,7 @@
 using PlatformPlatform.AccountManagement.Application;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using PlatformPlatform.AccountManagement.WebApi;
-using PlatformPlatform.AccountManagement.WebApi.Endpoints;
+using PlatformPlatform.AccountManagement.WebApi.Tenants;
 using PlatformPlatform.Foundation.AspNetCoreUtils;
 
 var builder = WebApplication.CreateBuilder(args);
